apps/server/requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-limiter==3.5.0
flask-talisman==1.1.0
gunicorn==23.0.0
flask-sqlalchemy==3.1.1
psycopg[binary]>=3.2
flask-migrate==4.0.7
PyJWT==2.8.0
resend==2.0.0
stripe==7.0.0
